fixed bugs

+++ Lab1C/lab1.c
@@ -31,7 +31,7 @@ void task1(Array* arr)
     printf("Элементы, большие суммы: ");
 
     for (size_t i = 0; i < array_size(arr); i++) {
-        if (array_get(arr, i) > sum) {
+        if (array_get((int)arr, i) > sum) {
             printf("%lu (index: %zu) ", array_get(arr, i), i);
         }
     }
